Separated ReactionData into its own submodule
