+++ src/scattering/photo22.rs
@@ -1,7 +1,9 @@
 use super::*;
 
 pub mod base;
+pub mod rx_data;
 pub mod cross_section;
 
 pub use base::*;
+pub use rx_data::*;
 pub use cross_section::*;
